+++ config/dependencies.global.php
@@ -1,6 +1,8 @@
 <?php
 
+use ZfrEbWorker\Cli\PublisherCommand;
 use ZfrEbWorker\Cli\WorkerCommand;
+use ZfrEbWorker\Container\PublisherCommandFactory;
 use ZfrEbWorker\Container\QueuePublisherFactory;
 use ZfrEbWorker\Container\WorkerCommandFactory;
 use ZfrEbWorker\Container\WorkerMiddlewareFactory;
@@ -15,6 +17,7 @@ return [
         ],
 
         'factories' => [
+            PublisherCommand::class => PublisherCommandFactory::class,
             QueuePublisher::class   => QueuePublisherFactory::class,
             WorkerCommand::class    => WorkerCommandFactory::class,
             WorkerMiddleware::class => WorkerMiddlewareFactory::class,